Set up SSH key

+++ setup.sh
@@ -8,6 +8,15 @@ fi
 # Install Command Line Tools if not yet installed.
 xcode-select -p 1>/dev/null || xcode-select --install
 
+# Set up SSH key if none exists
+if [[ ! -f ~/.ssh/id_rsa ]]; then
+  ssh-keygen -t rsa -b 4096 -C "ideler.dennis@gmail.com"
+  eval "$(ssh-agent -s)"
+  ssh-add ~/.ssh/id_rsa
+  pbcopy < ~/.ssh/id_rsa.pub
+  read -p "Add copied SSH key to https://github.com/settings/keys then press Enter to continue..."
+fi
+
 mkdir -p ~/github
 
 rm -rf ~/github/dideler/dotfiles
